Move article fetchers to article package
